requirements.txt
Django>=2.1
psycopg2==2.7.6.1
pycryptodome>=3.7.0
pytz==2018.7
pytest==3.10.1  # https://github.com/pytest-dev/pytest
pytest-sugar==0.9.2  # https://github.com/Frozenball/pytest-sugar
pytest-django==3.4.4  # https://github.com/pytest-dev/pytest-django
coverage==4.5.2  # https://github.com/nedbat/coveragepy
django-coverage-plugin==1.6.0  # https://github.com/nedbat/django_coverage_plugin
tox==3.7.0
black==20.8b1
djangorestframework==3.11.1